# docs/merchant/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/merchant/api/v1/merchant/": {
            "post": {
                "description": "Creates a new merchant using provided data and returns the merchant details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Create a new merchant",
                "parameters": [
                    {
                        "description": "Merchant creation payload",
                        "name": "merchant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Merchant created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantCreateResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request or error during creation",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/merchant/me/": {
            "get": {
                "description": "Retrieves details of the merchant associated with the current user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Get merchant details",
                "responses": {
                    "200": {
                        "description": "Merchant details",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantDetailResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Merchant not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the merchant information for the current user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant"
                ],
                "summary": "Update merchant information",
                "parameters": [
                    {
                        "description": "Merchant update payload",
                        "name": "merchant",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Merchant updated successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantCreateResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/merchant/me/api-key/": {
            "get": {
                "description": "Retrieves all API keys associated with the current merchant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant",
                    "APIKey"
                ],
                "summary": "List merchant API keys",
                "responses": {
                    "200": {
                        "description": "List of API keys",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantAPIKeyResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Merchant not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Generates and stores a new API key for the current merchant. The raw API key is returned only once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant",
                    "APIKey"
                ],
                "summary": "Create a new API key for a merchant",
                "parameters": [
                    {
                        "description": "API key creation payload",
                        "name": "apiKey",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.CreateAPIKeyRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "API key created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.CreateAPIKeyResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/merchant/me/api-key/{id}/": {
            "post": {
                "description": "Deactivates a merchant's API key specified by its ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant",
                    "APIKey"
                ],
                "summary": "Deactivate an API key",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "API key deactivated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid API key ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Error deactivating API key",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/merchant/api/v1/merchant/me/tokens/": {
            "get": {
                "description": "Retrieves a list of tokens associated with the current merchant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant",
                    "Token"
                ],
                "summary": "List merchant tokens",
                "responses": {
                    "200": {
                        "description": "List of merchant tokens",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantTokenCreateResponseDTO"
                            }
                        }
                    },
                    "404": {
                        "description": "Merchant not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new token for the current merchant.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Merchant",
                    "Token"
                ],
                "summary": "Create a merchant token",
                "parameters": [
                    {
                        "description": "Merchant token creation payload",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantTokenCreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token created successfully",
                        "schema": {
                            "$ref": "#/definitions/github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantTokenCreateResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid request or merchant not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.CreateAPIKeyRequestDTO": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.CreateAPIKeyResponseDTO": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantAPIKeyResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantCreateRequestDTO": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantCreateResponseDTO": {
            "type": "object",
            "properties": {
                "commision_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantDetailResponseDTO": {
            "type": "object",
            "properties": {
                "commision_rate": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantTokenCreateRequestDTO": {
            "type": "object",
            "required": [
                "active",
                "token_id"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "token_id": {
                    "type": "string"
                }
            }
        },
        "github_com_1stpay_1stpay_internal_transport_rest_merchant_restdto.MerchantTokenCreateResponseDTO": {
            "type": "object",
            "required": [
                "active",
                "created_at",
                "id",
                "merchant_id",
                "token_id"
            ],
            "properties": {
                "active": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                }
            }
        },
...
    }
}